参数化

+++ System/PID.h
@@ -1,6 +1,7 @@
 #ifndef _PID__H__
 #define _PID__H__
 
+#include <stdint.h>
 typedef struct
 {
 	float Target;
@@ -17,7 +18,7 @@ typedef struct
 }PID_Controller;
 
 void PID_Init(void);
-float PID_Calculate(PID_Controller *PID,float dt);
+float PID_Calculate(PID_Controller *PID,uint8_t dt);
 void PID_Set_Paramaters(PID_Controller *PID,float Kp,float Ki,float Kd,float Int_Limit,float Out_Limit);
 
 extern PID_Controller PID_Rate_Pitch;
